requirements.txt
fastapi
uvicorn
ephem
httpx
python-dotenv
pydantic